SI to consider only 'active' images for retention
Fixes rm2400

As mentioned in rm2400, SI currently doesn't take image status into
consideration while accounting for retention. As a result, SI is/maybe
deleting images that do not qualify for deletion. To overcome this, snapshot
worker is made to look only for images that are active while processing
retention.